// src/interfaces/interfacesApp.tsx
export interface DatosInicioSesion{
    nombreUsuario: string;
    password: string
}

export interface UsuarioLogueado{
    token: string;
    rol: string;
    nombre: string;
    apellidos: string;
    nombreUsuario: string;
    nombrePtoEmision: string;
    codEstablecimiento: string;
    codPtoEmision: string;
    ptoEmisionId: number;
}

export interface NuevoUsuario {
    nombre: string;
    apellido: string;
    nombreUsuario: string;
    password: string;
    ptoEmisionId: number;
}

export interface PtoEmision {
    id?:                 number;
    nombre:             string;
    codEstablecimiento: string;
    codPtoEmision:      string;
    secuencial:         number;
}

export interface DatoUsuario {
    id: number;
    nombre: string;
    apellido: string;
    nombreUsuario: string;
    ptoEmision?: PtoEmision;
    rol: Rol;

}

interface Rol{
    nombreRol: string;
}

export interface Factura {
    id:                 number;
    ambiente:           string;
    claveAcceso:        string;
    fechaEmision:       string;
    nombreCompletoCliente:     string;
    apellidos:          string;
    identificacion:     string;
    direccion:          string;
    placa:              string;
    email:              string;
    establecimiento:    string;
    ptoEmision:         string;
    secuencial:         string;
    fechaCreacion:      string;
    totalSinImpuestos:  number;
    subtotal12:         number;
    subtotal0:          number;
    totalIva:           number;
    totalDescuento:     number;
    total:              number;
    creadaPor:          string;
    detalles:           Detalle[];
    formasPago:         FormasPago[];
}


export interface Detalle {
    id:             number;
    codigoProducto: string;
    nombreProducto: string;
    cantidad:       number;
    precioUnitario: number;
    descuento:      number;
    porcentajeIva:  number;
    iva:            number;
    subtotal:       number;
}

export interface FormasPago {
    id:           number;
    codigo:       string;
    descripcion:  string;
    plazo:        null;
    unidadTiempo: null;
    valor:        number;
}


// Generated by https://quicktype.io

export interface NuevaFactura {
    fechaEmision:       Date;
    nombreCompletoCliente:     string;
    identificacion:     string;
    direccion:          string;
    email?:              string;
    placa?:              string;
    detalles:           DetalleNuevaFactura[];
    formasPago:         FormasPagoNuevaFactura[];
}

export interface DetalleNuevaFactura {
    uid: number
    valorAVender: number
    codigoProducto: string;
    nombreProducto: string;
    cantidad:       number;
    precioUnitario: number;
    descuento:      number;
    porcentajeIva:  number;
}

export interface FormasPagoNuevaFactura {
    uid: number;
    codigo:      string;
    descripcion: string;
    valor:       number;
}

export interface TotalesNuevaFactura{
    totalSinImpuestos:  number;
    subtotal12:         number;
    subtotal0:          number;
    totalIva:           number;
    totalDescuento:     number;
    total:              number;
}

export interface ClienteNuevaFactura{
    nombreCompletoCliente:     string;
    identificacion:            string;
    direccion:                 string;
    email?:                    string;
    placa?:                     string;
}

export interface Filtro {
    busqueda:  string;
}

export interface Producto {
    uid: number
    codigoProducto: string;
    nombreProducto: string;
    tieneIva:       boolean;
    precioUnitario: number;
    porcentajeIva:  number;
}

export interface CatalogoFormaPago {
    codigo:      string;
    descripcion: string;
}